updater can now uninstall - next commit will be focusing on elevation and cleanup - pre4 coming after cleanup
